playbook-agent/requirements.txt
pyyaml>=6.0
firebase-admin>=6.0.0
google-cloud-storage>=2.0.0
kubernetes>=28.0.0
